Implement constant references in compiler IL (initially disabled)
This commit introduces a representation for non-null reference constants
in the IL. For each known object, there is now a lazily created symbol
reference representing the constant that always refers to that object.
The symref has a known object index in the same way that existing
"improved" symrefs do. It's static, and the IL should load it directly,
i.e. using aload. A constant represented as such a load looks as follows
in tracing:

    aload  <constant object ref>[#429  final Static] (obj3)

This new representation is meant to completely supplant the existing
ways of attaching known object information to nodes in the IL, namely
the per-node known object index and the aforementioned "improved"
symrefs. There are two reasons for this change:

1. These reference constants are inherently guaranteed to evaluate to
   references to the expected objects, unlike nodes where we've used the
   prior techniques, since those left the original operations in place.
   This guarantee makes it less error prone to constant-fold loads, and
   it makes it possible to do so in more circumstances.

2. These reference constants make the resulting references available
   more or less directly. It generally takes only a single load to get
   the reference. This improves performance in cases where the reference
   was originally coming from a chain of two or more dependent loads
   (and where it doesn't become unused before code generation).

The new preprocessor symbol TR_ALLOW_NON_CONST_KNOWN_OBJECTS gates old
functionality that should no longer be used once const refs are enabled.
In particular, if that symbol is not defined, then attempts to get or
set the per-node known object index, and attempts to create an
"improved" version of an existing symref, will cause a build error. This
is useful for making sure that every use of those APIs has alternative
logic implemented to use const refs instead (when enabled). It should
also eventually be useful for eliminating those old APIs. However, it's
expected that there will be builds that need to disable const refs for
some time yet. Indeed, this commit does not even enable const refs by
default.

This commit also implements a cross-platform code generation strategy
for const refs, along with minimal architecture-specific instruction
selection changes. During evaluation, each known object used as a const
ref will be assigned a label, and the generated instructions will use
the label to identify the address from which to load. Later, the
reference slots will be allocated amongst the snippets in the binary
code, in particular between the regular snippets and the data snippets.
Before allocating the slots though, the downstream project needs to add
labels for any necessary keepalives (see RetainedMethodSet), and then it
has an opportunity to specify any arbitrary order for the references to
appear in, as convenient for the downstream implementation. Finally, at
the end of the compilation, the downstream project must arrange for the
slots to be scanned by GC as appropriate, and it must initialize them to
point to their respective objects.